% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_pos_any_neg.R
\name{any_pos_any_neg}
\alias{any_pos_any_neg}
\title{Find any positve gven a list of column}
\usage{
any_pos_any_neg(
  df,
  test_cols,
  test_type,
  neg_value = "No",
  pos_value = "Yes",
  id_cols = c("Participant_PPID")
)
}
\arguments{
\item{df}{A data.frame or data.table or tibble}

\item{test_cols}{a vector of column names eg cols measuring symptoms}

\item{test_type}{vector of length one to be appended to output name}

\item{neg_value}{a value that shows  the negative outcome in test_cols}

\item{pos_value}{a value that measures the positive outcome in test_cols}

\item{id_cols}{A vector of column names in df that you would want included eg participant ID}
}
\value{
A data.table
}
\description{
Given a list of column names eg symptoms find if there are any positives
}
\examples{
data("tb_resp_symptoms")

respirotory_symptoms <- c("COUGH", "EXPECTORATION",
                          "HEMOPTYSIS", "CHEST_PAIN",
                          "DYSPNOE")

any_pos_any_neg(df = tb_resp_symptoms, 
                 test_cols =respirotory_symptoms ,
                 test_type = "resp_symp" ,
                 neg_value ="No",
                 pos_value = "Yes",
                 id_cols ="ID")
}
