% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_os.R
\name{auth_os}
\alias{auth_os}
\title{Authenticate to the OpenSpecimen API}
\usage{
auth_os(url, username, password, domain_name = "openspecimen")
}
\arguments{
\item{url}{The URL of the OpenSpecimen API.}

\item{username}{The username for authentication.}

\item{password}{The password for authentication.}

\item{domain_name}{The domain name (default is "openspecimen").}
}
\value{
A list with the URL and the authentication response which contains auth token and user details
}
\description{
This function authenticates a user to the OpenSpecimen API using the provided
URL, username, and password.
}
\note{
See additional information on this API on https://openspecimen.atlassian.net/wiki/spaces/CAT/pages/1116035/REST+APIs
}
\examples{

# os_test_url <- Sys.getenv("OSTESTURL")
# os_test_username <- Sys.getenv("OSUSERNAME")
# os_test_password <- Sys.getenv("OSPASSWORDTEST")
# 
# auth_repo <- auth_os(url = os_test_url,
#                      username = os_test_username,
#                      password = os_test_password)

}
