% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_os_df_to_wide.R
\name{convert_os_df_to_wide}
\alias{convert_os_df_to_wide}
\title{Title Convert  OS data to wide
Description Convert OS data to wide and merge with clinical data}
\usage{
convert_os_df_to_wide(
  clinical_df,
  specimen_df,
  join_by_col,
  specimen_col,
  specimen_label_col,
  clinical_vars = NULL,
  os_aliquot_names,
  join_function = "left_join"
)
}
\arguments{
\item{clinical_df}{A data frame with clinical data at Find typically from Open Clinica.}

\item{specimen_df}{A data frame with specimen data at Find generally Open Specimen.}

\item{join_by_col}{A column used to merge clinical data and specimen data}

\item{specimen_col}{The column containing specimen names.
At find usually Specimen Requirement name}

\item{specimen_label_col}{Column containing specimen IDS in Open Specimen this is
specimen label}

\item{clinical_vars}{A vector of clinical variables.
Ideally variables that usually describe patient characteristics eg sex,
disease information etc}

\item{os_aliquot_names}{A vector of the aliquot names that you want to use.}

\item{join_function}{A function in char to join with. Recommended to use
dplyr merge function eg left_join, inner_join etc. Recommended to load dplyr}
}
\value{
A wide data frame with one row per participant and one column for
each aliquot name and the number of aliquots and the specimen labels for
that aliquot.
}
\description{
Title Convert  OS data to wide
Description Convert OS data to wide and merge with clinical data
}
\examples{
library(data.table)
library(dplyr)
data("tb_data", package = "findBiobankR")
data("tb_specimen_df", package = "findBiobankR")

# required clinical cols
clinical_cols <- c("tb_group", "hiv_status", "sex", "country") 

clin_specs  <- convert_os_df_to_wide(clinical_df = tb_data,
                                     specimen_df = tb_specimen_df, 
                                     join_by_col = "ppid",
                                     specimen_col ="specimen_type" ,
                                     clinical_vars = clinical_cols ,
                                     specimen_label_col = "specimen_label",
                                     os_aliquot_names = c("Serum", "Plasma")
)

knitr::kable(clin_specs[1:10, ])
}
