% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folders}
\alias{create_folders}
\title{Create Project Folders}
\usage{
create_folders(
  main_folder,
  subfolders = c("data", "results", "doc"),
  suffix = "_folder"
)
}
\arguments{
\item{main_folder}{A string specifying the path of the main project folder.}

\item{subfolders}{A character vector of subfolders to be created within the main folder.
Default is c("data", "results", "doc").}

\item{suffix}{A string to append to each subfolder name when creating global variables.
Default is "_folder".}
}
\description{
This function creates a set of folders for a project, including a main folder and specified subfolders.
It optionally assigns global variable names for each folder path, using a provided suffix.
}
\examples{
path_name <- tempfile(pattern = "main_folder") \%>\%
    normalizePath(winslash = "/")

create_folders(path_name)

# Specify different subfolders and a custom suffix
create_folders(path_name, c("archive", "temp", "shared"), "_dir")

}
