% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table.R
\name{data_table}
\alias{data_table}
\title{data_table function}
\usage{
data_table(df, pageLength = 50, row_names = FALSE, file_name = "output_file")
}
\arguments{
\item{df}{The data frame to be displayed as a datatable.}

\item{pageLength}{The number of rows to display per page. Default is 50.}

\item{row_names}{Logical value indicating whether to display row names. Default is FALSE.}

\item{file_name}{The name of the file to be used for exporting data.}
}
\value{
An interactive data table of class DT
}
\description{
This function creates a datatable from a data frame using the DT package.
}
\examples{
 data("iris")
df = iris[1:10,]

data_table(df, 
           file_name = "iris_sub_df")
}
