% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBiobankR-package.R
\docType{package}
\name{findBiobankR-package}
\alias{findBiobankR}
\alias{findBiobankR-package}
\title{findBiobankR: Find Biobank R Package}
\description{
A collection of functions used by Find Biobank team.
}
\author{
\strong{Maintainer}: Moses Mburu \email{moses.mburu@finddx.org} (\href{https://orcid.org/0000-0001-7281-583X}{ORCID})

}
\keyword{internal}
