% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aliquot_number.R
\name{get_aliquot_number}
\alias{get_aliquot_number}
\title{Get the aliquot number from a aliquot ID.}
\usage{
get_aliquot_number(x)
}
\arguments{
\item{x}{A character vector representing aliquot IDs.}
}
\value{
A character vector of aliquot numbers. If the input does not have 16 characters, returns NA with a warning.
}
\description{
This function extracts the aliquot number from a aliquot ID.
}
\examples{

get_aliquot_number("PP0070199910001001")


}
