% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bulk_order_detail.R
\name{get_bulk_order_detail}
\alias{get_bulk_order_detail}
\title{Get Bulk OpenSpecimen Order Details}
\usage{
get_bulk_order_detail(auth_response, orders_ids, ...)
}
\arguments{
\item{auth_response}{The authentication response obtained from \code{auth_os}.}

\item{orders_ids}{A vector of order IDs for which details are to be retrieved.}

\item{...}{Additional parameters to be passed to \code{get_order_detail}.}
}
\value{
A data.table containing the parsed details for the specified orders.
}
\description{
This function retrieves details for a bulk list of orders from the OpenSpecimen API using the provided authentication response and order IDs.
}
\examples{
#get_bulk_order_detail()
}
