% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bulk_orders.R
\name{get_bulk_orders}
\alias{get_bulk_orders}
\title{Get OpenSpecimen Orders in Bulk}
\usage{
get_bulk_orders(
  auth_response,
  include_stats = TRUE,
  page_size = 250,
  max_pages = 10000
)
}
\arguments{
\item{auth_response}{The authentication response obtained from \code{auth_os}.}

\item{include_stats}{Logical, whether to include statistics (default is TRUE).}

\item{page_size}{Integer, the number of results to retrieve per page (default is 250).}

\item{max_pages}{Integer, the maximum number of pages to retrieve (default is 10,000).}
}
\value{
A data frame containing the aggregated order data from multiple pages.
}
\description{
This function retrieves a list of orders from the OpenSpecimen API in bulk, handling pagination and combining results. It uses the provided authentication response and optional query parameters for customization.
}
\note{
This function allows you to efficiently retrieve orders from OpenSpecimen in bulk while handling pagination and result aggregation.
}
\examples{
#get_bulk_orders()
}
