% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bulk_query.R
\name{get_bulk_query}
\alias{get_bulk_query}
\title{Execute Bulk Query}
\usage{
get_bulk_query(
  auth_response,
  query_id,
  driving_form = "Specimen",
  wide_row_mode = "OFF",
  page_size = 5000,
  max_pages = 10000
)
}
\arguments{
\item{auth_response}{The authentication response obtained from \code{auth_os}.}

\item{query_id}{The saved query identifier.}

\item{driving_form}{The driving form for the query (default is "Specimen").}

\item{wide_row_mode}{Specifies whether multi-valued attributes result in a single row or one row per value (default is "OFF"). Other permitted values are SHALLOW and DEEP.}

\item{page_size}{Page size for each query.}

\item{max_pages}{Maximum number of pages to fetch.}
}
\value{
A data table containing the results of the bulk query.
}
\description{
Execute the "query" method multiple times to fetch all data by page.
}
\note{
Make sure to provide an HTTP response with the appropriate structure for parsing.
}
\examples{


# auth_response <- auth_os(url =  Sys.getenv("OSTESTURL"),
#                          username = Sys.getenv("OSUSERNAME"),
#                          password = Sys.getenv("OSPASSWORDTEST"))
# 
# 
# cv_samples <- get_bulk_query(auth_response, query_id = 105)
#   
# 
# cv_samples[1:5]
}
