% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_id.R
\name{get_event_id}
\alias{get_event_id}
\title{Get the event ID from a Aliquot ID.}
\usage{
get_event_id(x)
}
\arguments{
\item{x}{A character vector representing Aliquot IDs.}
}
\value{
A character vector of event IDs.
}
\description{
This function extracts the event ID from a Aliquot ID.
}
\examples{

get_event_id("PP0070199910001001")

}
