% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_find_study_id.R
\name{get_find_study_id}
\alias{get_find_study_id}
\title{Extract FinDx Study ID}
\usage{
get_find_study_id(x)
}
\arguments{
\item{x}{The input string from which to extract the study ID.}
}
\value{
A character vector containing the extracted study ID.
}
\description{
This function extracts a study ID from participant or aliquot ID
}
\examples{

ppid <- c("TB002133530", "TB032020154",
          "TB001200533", "TB001035037", 
          "TB002133927")
study_id <- get_find_study_id(ppid)

}
