% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_order_items.R
\name{get_order_items}
\alias{get_order_items}
\title{Retrieve Order Items from OpenSpecimen}
\usage{
get_order_items(auth_response, order_id)
}
\arguments{
\item{auth_response}{Authentication response from OpenSpecimen.}

\item{order_id}{ID of the order whose items are requested.}
}
\value{
A data.table containing parsed information from the OpenSpecimen order items.
}
\description{
This function retrieves the details of order items of the requested order from OpenSpecimen.
}
\examples{
#get_order_items()
}
