% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orders.R
\name{get_orders}
\alias{get_orders}
\title{Get Orders}
\usage{
get_orders(
  auth_response,
  include_stats = TRUE,
  start_at = NULL,
  max_results = NULL
)
}
\arguments{
\item{auth_response}{Authentication response object obtained from the OpenSpecimen API.}

\item{include_stats}{If TRUE, includes statistics in the response (optional, default is TRUE).}

\item{start_at}{Starting row of the result (optional).}

\item{max_results}{Maximum number of records to fetch (optional).}
}
\value{
A data frame containing the details of distribution orders, or NULL if an error occurs.
}
\description{
Use this function to retrieve distribution orders in the OpenSpecimen application.
The function sends an HTTP GET request to the OpenSpecimen server to fetch details of distribution orders.
}
\examples{
\dontrun{
# Example usage:
auth_response <- authenticate_user(username = "your_username", password = "your_password")
orders <- get_orders(auth_response)
}

#get_orders()
}
\seealso{
\code{\link{auth_os}}, \code{\link{parse_os_response}}, \code{\link{parse_os_order_data}}
}
