% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os_query.R
\name{get_os_query}
\alias{get_os_query}
\title{Execute a Saved Query}
\usage{
get_os_query(
  auth_response,
  query_id,
  driving_form = c("Specimen", "Participant"),
  wide_row_mode = c("OFF", "DEEP", "SHALLOW"),
  start_at = NULL,
  max_results = NULL
)
}
\arguments{
\item{auth_response}{The authentication response obtained from \code{auth_os}.}

\item{query_id}{The ID of the saved query to execute.}

\item{driving_form}{The driving form for the query (default is "Specimen").}

\item{wide_row_mode}{The wide row mode for the query (default is "OFF").}

\item{start_at}{The starting position for the query (optional).}

\item{max_results}{The maximum number of results to retrieve (optional).}
}
\value{
A data frame containing the results of the saved query.
}
\description{
This function executes a saved query in the OpenSpecimen API using the provided
authentication response and query ID. You can specify the parameters for the query
such as driving form, wide row mode, start position, and maximum results.
}
\note{
See additional information on this API on https://openspecimen.atlassian.net/wiki/spaces/CAT/pages/104529939/Query
}
\examples{

# os_test_url <- Sys.getenv("OSTESTURL")
# os_test_username <- Sys.getenv("OSUSERNAME")
# os_test_password <- Sys.getenv("OSPASSWORDTEST")
# 
# auth_repo <- auth_os(url = os_test_url,
#                      username = os_test_username,
#                      password = os_test_password)
# 
# get_os_query(auth_repo, query_id = 114,
#              driving_form = "Specimen",
#              wide_row_mode = "OFF",
#              start_at = 0,
#              max_results = 5)
}
