% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os_site.R
\name{get_os_site}
\alias{get_os_site}
\title{Get OpenSpecimen Site}
\usage{
get_os_site(auth_response, site_id)
}
\arguments{
\item{auth_response}{Authentication response object obtained from the OpenSpecimen API.}

\item{site_id}{The unique identifier of the site to fetch.}
}
\value{
A data frame containing the details of the specified site.
}
\description{
Use this function to retrieve details of a site from the OpenSpecimen application.
}
\examples{
#get_os_site()
}
