% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_id.R
\name{get_site_id}
\alias{get_site_id}
\title{Get the site ID from a participant ID in a clinical study.}
\usage{
get_site_id(x)
}
\arguments{
\item{x}{A character vector containing participant IDs.}
}
\value{
A character vector containing the site IDs.
}
\description{
This function extracts the site ID from a participant ID in a FindDX clinical study.
}
\examples{
get_site_id("PP007019990")
}
