% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specimen_type_id.R
\name{get_specimen_type_id}
\alias{get_specimen_type_id}
\title{Get the specimen type ID from a Aliquot ID.}
\usage{
get_specimen_type_id(x)
}
\arguments{
\item{x}{A character vector representing Aliquot IDs.}
}
\value{
A character vector of specimen type IDs.
}
\description{
This function extracts the specimen type ID from a Aliquot ID.
}
\examples{

get_specimen_type_id("PP0070199910001001")

}
