% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_form_name.R
\name{groupby_vec}
\alias{groupby_vec}
\title{Group By Vector Elements and Count Frequencies}
\usage{
groupby_vec(df, myvec, byvec)
}
\arguments{
\item{df}{The input data frame.}

\item{myvec}{A vector of column names to be used for grouping and counting frequencies.}

\item{byvec}{A vector of column names for grouping the data frame. ID columns that don't change}
}
\value{
A data table summarizing the frequencies of non-empty and non-missing values in
each group for the specified columns.
}
\description{
This function takes a data frame, a vector of column names (\code{myvec}), and a vector of grouping
column names (\code{byvec}). It then groups the data frame by each element in \code{myvec} along with the
specified grouping columns (\code{byvec}). For each group, it counts the frequencies of non-empty
and non-missing values in the specified column (\code{myvec_i}) and returns a summary data frame.
}
\examples{
#groupby_vec()
}
