% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clean_os_names.R
\name{make_clean_os_names}
\alias{make_clean_os_names}
\title{Clean Open Specimen Names}
\usage{
make_clean_os_names(names_os_df)
}
\arguments{
\item{names_os_df}{column names from Open specimen data or Open Specimen data.frame}
}
\value{
A character vector of names or data.frame
}
\description{
Data set from Open specimen normally come with very programming unfriendly column
names with repeat suffixes # and so on this function attempts to correct this
}
\examples{

nms_os <- c("Participant_PPID", "Visit_Visit Site",
            "Visit_Name", "Visit_Event Label", 
            "Specimen_Specimen Label",
            "Specimen_Type", "Specimen_Lineage", "Specimen_Barcode", 
            "Specimen_Container Name", "Specimen_Container Position", 
            "Specimen_Requirement Name", "Collection Protocol_Short Title")

make_clean_os_names(nms_os)
}
