% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_all_duplicates.R
\name{mark_all_duplicates}
\alias{mark_all_duplicates}
\title{Mark All Duplicates in a Vector}
\usage{
mark_all_duplicates(vec)
}
\arguments{
\item{vec}{A numeric or character vector where duplicates are to be identified.}
}
\value{
A logical vector of the same length as \code{vec}, where TRUE indicates that
the corresponding element in \code{vec} is a duplicate.
}
\description{
This function identifies all duplicate elements in a vector, marking both the first
occurrence and subsequent duplicates.
}
\examples{

vec <- c("apple", "banana", "apple", "cherry", "banana")
mark_all_duplicates(vec)
}
