% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_thousandth_sep.R
\name{num_thousandth_sep}
\alias{num_thousandth_sep}
\title{Format Numeric Values with Thousands Separator}
\usage{
num_thousandth_sep(x, sep = ",")
}
\arguments{
\item{x}{A numeric atomic vector to be formatted with a thousands separator.}

\item{sep}{The character used as the thousands separator (default is a comma).}
}
\value{
A character vector with formatted numeric values.
}
\description{
This function takes a numeric atomic vector \code{x} and formats the values with a thousands separator using \code{scales::label_comma()}. If the input \code{x} is not a numeric atomic vector, it will throw an error.
}
\note{
This function is useful for formatting numeric values in a human-readable format.
}
\examples{

num_thousandth_sep(100)
num_thousandth_sep(1234567)
num_thousandth_sep(c(1234, 5678, 9876))

}
