% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_all_sites.R
\name{parse_all_sites}
\alias{parse_all_sites}
\title{Parse All Sites}
\usage{
parse_all_sites(response)
}
\arguments{
\item{response}{The response object obtained from the OpenSpecimen API.}
}
\value{
A data frame containing the details of all sites.
}
\description{
Use this function to parse the response content from the OpenSpecimen API
and extract the details of all sites.
}
\examples{
#parse_all_sites()
}
