% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{parse_data}
\alias{parse_data}
\title{Parse Response Data}
\usage{
parse_data(response)
}
\arguments{
\item{response}{The response data to be parsed.}
}
\value{
A data frame containing the parsed data.
}
\description{
This function parses the response data obtained from the OpenSpecimen API and
organizes it into a data frame.
}
\note{
Make sure to provide an HTTP response with the appropriate structure for parsing.
}
\examples{
#parse_data()
}
