% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_order_detail_data.R
\name{parse_order_detail_data}
\alias{parse_order_detail_data}
\title{Parse OpenSpecimen Order Detail Data}
\usage{
parse_order_detail_data(response, remove_personal_info = TRUE)
}
\arguments{
\item{response}{The response object from the Open Specimen API.}

\item{remove_personal_info}{Logical, whether to remove personal information columns (default is TRUE).}
}
\value{
A data.table containing the parsed order detail data.
}
\description{
This function parses the response content from the OpenSpecimen API for order details and organizes it into a data.table.
}
\examples{

# order_detail <- parse_order_detail_data(response)

}
