% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_os_order_data.R
\name{parse_os_order_data}
\alias{parse_os_order_data}
\title{Parse OpenSpecimen Order Data}
\usage{
parse_os_order_data(response, remove_personal_info = TRUE)
}
\arguments{
\item{response}{The response data to be parsed.}

\item{remove_personal_info}{Logical, whether to remove personal information columns (default is TRUE).}
}
\value{
A data table containing the parsed order data.
}
\description{
This function parses the response data obtained from the OpenSpecimen API and organizes it into a data frame. It assumes that the response is in a structured format suitable for conversion to a data frame.
}
\note{
See API documentation for details on the response structure. https://openspecimen.atlassian.net/wiki/spaces/CAT/pages/259457077/Retrieve+list+of+Orders
}
\examples{
#parse_os_order_data()
}
