% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_os_response.R
\name{parse_os_response}
\alias{parse_os_response}
\title{OpenSpecimen API Request Response Handler}
\usage{
parse_os_response(response, parse_data_function)
}
\arguments{
\item{response}{The response object from the OpenSpecimen API request.}

\item{parse_data_function}{The function used to parse the response data}
}
\value{
Parsed content from the API response.
}
\description{
This function handles the response from an OpenSpecimen API request, checks the status code,
and returns the parsed response content or handles errors as needed.
}
\examples{
#parse_os_response()
}
