% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_site_details.R
\name{parse_site_details}
\alias{parse_site_details}
\title{Parse OpenSpecimen Site Details}
\usage{
parse_site_details(response)
}
\arguments{
\item{response}{The API response object.}
}
\value{
A data table containing the parsed details of the site.
}
\description{
Use this function to parse the response and extract details of a site from the OpenSpecimen application.
}
\examples{
#parse_site_details()
}
