% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multiple_files.R
\name{read_multiple_files}
\alias{read_multiple_files}
\title{Read multiple files from a folder}
\usage{
read_multiple_files(
  folder_path,
  read_function = "fread",
  pattern = NULL,
  file_names = NULL,
  ...
)
}
\arguments{
\item{folder_path}{path to file}

\item{read_function}{The read function eg read.csv,
read_excel, fread. Depends on your file types or function preference.
Please note to load the library you want to use the function from}

\item{pattern}{A pattern to be passed to list.files function}

\item{file_names}{If you have a vector of file names then this could also be used instead of pattern}

\item{...}{Other parameters from read function used}
}
\value{
returns a list of data frames

A list of data frames
}
\description{
Read Files From excel workbook
}
\examples{
iris_species <- system.file("extdata", "iris_species.xlsx",
                          package = "findBiobankR")

iris_list <- read_multiple_sheets(file_path  = iris_species ) \%>\% 
  lapply(head, 4)
iris_list

library(data.table)
iris_split <- split(iris, 
                    f = iris$Species)


path_name <- tempfile(pattern = "read_multipe_files") \%>\%
    normalizePath(winslash = "/")

dir.create(path_name)
file_names <- paste0(names(iris_split),  ".csv")
# create file name
file_paths <- file.path(path_name, file_names)
lapply(seq_along(file_paths), function(x){
    df = iris_split[[x]]
    fwrite(df, file = file_paths[x])
})

read_multiple_files(folder_path =path_name ,
                    read_function = "fread",
                    pattern = ".csv$",
                    file_names =NULL)
#file.remove(file_name)
unlink(path_name)
}
