% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_os_data.R
\name{read_os_data}
\alias{read_os_data}
\title{Read Open Specimen data.frame in R}
\usage{
read_os_data(path_name, header_true = FALSE, ...)
}
\arguments{
\item{path_name}{Path to OS data}

\item{header_true}{Whether the file has a header defaults to false}

\item{...}{Other parameters taken by data.table::fread}
}
\value{
data.table
}
\description{
A function to read Open Specimen data. The need comes from the fact that this function standadizes names
}
\examples{
tb_specs <- system.file("extdata", "tb_specs.csv",
                          package = "findBiobankR")
read_os_data(path_name = tb_specs)
}
