% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_all_na_col.R
\name{rm_all_na_col}
\alias{rm_all_na_col}
\title{Removes columns that de selects/deletes  columns with missing rows}
\usage{
rm_all_na_col(df, missing_perc = 99.99, na_chars = c("", "NA", " "))
}
\arguments{
\item{df}{data.frame}

\item{missing_perc}{threshold to delete columns eg delete with more than 99 \% missing number should be between 0 and 100 \%}

\item{na_chars}{character vectors representing NA's}
}
\value{
A data.table
}
\description{
Given a data.frame, this function will remove columns with missing values above a certain threshold.
}
\examples{
library(data.table)

# Create a data frame with missing values.
df <- data.table(
  id = 1:10,
  name = c("John", "Jane", "", "Mary", "Mike", "", "", "Peter", "Sam", "Susan"),
  age = c(20, 21, 18, 22, 23, 19, NA, 24, 25, 26),
  height = c(170, 160, 180, 175, 185, 175, NA, 190, 195, 200),
  plasma_collected = rep(NA, 10)
)

# Remove columns with more than 99\% missing values.
rm_all_na_col(df, missing_perc = 99.99)
}
