% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_form_name.R
\name{rm_form_name}
\alias{rm_form_name}
\title{Remove Form Name Prefix from Data Frame Column Names}
\usage{
rm_form_name(df, form_name = "participant_additional_fields_tb_c_ps_")
}
\arguments{
\item{df}{Data frame whose column names need to be modified.}

\item{form_name}{Prefix to be removed from the column names. Default is "participant_additional_fields_tb_c_ps_".}
}
\value{
A modified data frame with the specified form name prefix removed from column names.
}
\description{
This function removes a specified form name prefix from the column names of a data frame. Usually from Open specimen
}
\examples{
\dontrun{
# Assuming df is a data frame with column names like "participant_additional_fields_tb_c_ps_Field1"
df_modified <- rm_form_name(df)
}

#rm_form_name()
}
