% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_distributed_summary.R
\name{samples_distributed_summary}
\alias{samples_distributed_summary}
\title{A summary of samples distributed}
\usage{
samples_distributed_summary(
  df,
  unique_by_col = "Participant_PPID",
  wider_by_col = "Specimen_Requirement Name",
  group_by_col = "tb_group",
  save_file = FALSE,
  path_name = NULL,
  order_col = "Total"
)
}
\arguments{
\item{df}{Clinical data with selected samples}

\item{unique_by_col}{Column name to calculate unique IDS}

\item{wider_by_col}{column to convert long data.frame to wide data.frame}

\item{group_by_col}{column names to group your data with and calculate frequencies}

\item{save_file}{whether to save your file}

\item{path_name}{File path to save your file}

\item{order_col}{Column to order your final summary}
}
\value{
data.frame
}
\description{
This functions calculates a summary of distributed samples
}
\examples{
library(data.table)
library(dplyr)
data("tb_data", package = "findBiobankR")
data("tb_specimen_df", package = "findBiobankR")
data("tb_dictionary", package = "findBiobankR")
# required clinical cols
clinical_cols <- c("tb_group", "hiv_status", "sex", "country") 

clin_specs  <- convert_os_df_to_wide(clinical_df = tb_data,
                                     specimen_df = tb_specimen_df, 
                                     join_by_col = "ppid",
                                     specimen_col ="specimen_type" ,
                                     clinical_vars = clinical_cols ,
                                     specimen_label_col = "specimen_label",
                                     os_aliquot_names = c("Serum", "Plasma")
)


clin_specs[1:4, serum_aliquot := select_one_aliquot(NP_labels_serum, aliquot_number = 1 )]
clin_specs[1:4, plasma_aliquot := select_one_aliquot(NP_labels_plasma, aliquot_number= 1 )]


selected_samples <- selected_aliquots_df(clin_specs,
                                         aliquot_cols = c("serum_aliquot","plasma_aliquot"),
                                         ppid_col = "ppid")

selected_samples_df <- process_save_selected_aliquots(file_path = NULL,
                               save_file= FALSE,
                               selected_samples = selected_samples,
                               os_position = tb_specimen_df,
                               ppid_col = "ppid",
                               specimen_id_col = "specimen_label",
                               os_required_cols = c("ppid",
                                                    "Specimen_Specimen Label",
                                                    "Specimen_Type",
                                                    "Specimen_Requirement Name", 
                                                    "Specimen_Barcode",
                                                    "Specimen_Container Name", 
                                                    "Specimen_Container Position",
                                                    "Scanned_barcode"))

list_dfs <- gen_clinical_data(clinical_data = tb_data,
                  clinical_cols= clinical_cols,
                  dictionary = tb_dictionary,
                  selected_samples_df = selected_samples_df,
                  ppid_col = "ppid",
                  dict_variable = "variable_name",
                  specimen_cols = c("ppid", "Specimen_Specimen Label", "Specimen_Type"),
                  additional_data = NULL )

samples_distributed_summary(df =list_dfs$clinical_data ,
         unique_by_col = "ppid",
         wider_by_col =  "Specimen_Type",
         group_by_col = "tb_group",
         save_file = FALSE,
         path_name = NULL,
         order_col = "Total")
}
