% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_one_aliquot.R
\name{select_one_aliquot}
\alias{select_one_aliquot}
\title{Title Select One aliquot}
\usage{
select_one_aliquot(x, aliquot_number = 1)
}
\arguments{
\item{x}{The input is a list column or a list of vectors}

\item{aliquot_number}{The number indicating the position of the desired aliquot label. Defaults to 1.}
}
\value{
A list of selected aliquot labels.

A vector
}
\description{
Description This function selects a specific aliquot label from a list of objects/ A list column
}
\examples{
library(data.table)
library(dplyr)
data("tb_data", package = "findBiobankR")
data("tb_specimen_df", package = "findBiobankR")

# required clinical cols
clinical_cols <- c("tb_group", "hiv_status", "sex", "country") 

clin_specs  <- convert_os_df_to_wide(clinical_df = tb_data,
                                     specimen_df = tb_specimen_df, 
                                     join_by_col = "ppid",
                                     specimen_col ="specimen_type" ,
                                     clinical_vars = clinical_cols ,
                                     specimen_label_col = "specimen_label",
                                     os_aliquot_names = c("Serum", "Plasma")
)

##
clin_specs[, select_one_aliquot(NP_labels_serum)]

# or data table way
clin_specs[, serum_aliquot_dt := select_one_aliquot(NP_labels_serum)]
# or tidyverse way
clin_specs \%>\%
  mutate(serum_aliquot_dp := select_one_aliquot(NP_labels_serum))

knitr::kable(clin_specs[1:10, ])

samples_list <- list(c("TB0990100010101001", "TB0990100010101002", "TB0990100010101003"),
                     c("TB0990100020101001", "TB0990100020101002", "TB0990100020101003"), 
                     c("TB0990100030101001", "TB0990100030101002", "TB0990100030101003"),
                     c("TB0990100040101001", "TB0990100040101002", "TB0990100040101003"),
                     c("TB0990100050101001", "TB0990100050101002", "TB0990100050101003"
                     ))
select_one_aliquot(samples_list, aliquot_number = 2)
}
