% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_many_aliquots.R
\name{select_patients_per_group}
\alias{select_patients_per_group}
\title{Select patients per group
For sample selections often this is based on patient groups and required number of patients maybe be different per patient group. This function helps with this selection}
\usage{
select_patients_per_group(
  df,
  patient_group_col,
  patient_groups,
  number_per_group,
  sort_cols
)
}
\arguments{
\item{df}{A data.frame with clinical data and number of aliquots per patient column}

\item{patient_group_col}{The column name with patient characterization}

\item{patient_groups}{A vector of patient group. This names should be categories in patient_group_col}

\item{number_per_group}{A vector same length indicating how many patient to select per group}

\item{sort_cols}{column(s) names that should be used to sort your data frame. We prefer to select patients with higher of aliquots}
}
\value{
data.table
}
\description{
Select patients per group
For sample selections often this is based on patient groups and required number of patients maybe be different per patient group. This function helps with this selection
}
\examples{

library(data.table)
library(dplyr)
data("tb_data", package = "findBiobankR")
data("tb_specimen_df", package = "findBiobankR")

# required clinical cols
clinical_cols <- c("tb_group", "hiv_status", "sex", "country") 

clin_specs  <- convert_os_df_to_wide(clinical_df = tb_data,
                                     specimen_df = tb_specimen_df, 
                                     join_by_col = "ppid",
                                     specimen_col ="specimen_type" ,
                                     clinical_vars = clinical_cols ,
                                     specimen_label_col = "specimen_label",
                                     os_aliquot_names = c("Serum", "Plasma")
)


## select 2 patients per group 
patient_groups  <- c("TB, Neg", "TB, Pos")
number_per_group  <- c(2, 2)
sort_cols = "NP_aliqN_serum"
clin_specs_2 <- select_patients_per_group(df = clin_specs, 
                                          patient_group_col = "tb_group",
                                          patient_groups = patient_groups,
                                          number_per_group = number_per_group,
                                          sort_cols = sort_cols)


}
