% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_aliquots_df.R
\name{selected_aliquots_df}
\alias{selected_aliquots_df}
\title{Create a data frame with selected aliquots in long format}
\usage{
selected_aliquots_df(
  clin_specs,
  aliquot_cols,
  ppid_col = "Participant_PPID",
  aliquot_name_col = "aliquot_name",
  spec_label_col = "specimen_label"
)
}
\arguments{
\item{clin_specs}{A data from with the selected aliquots as column}

\item{aliquot_cols}{Column names of the selected aliquots in \code{clin_specs} data frame}

\item{ppid_col}{Participant ID column}

\item{aliquot_name_col}{The name of the variable name after data.table::melt}

\item{spec_label_col}{The name of the value name after data.table::melt}
}
\value{
a data.table with \verb{ppid, specimen_label_col} columns
}
\description{
This function is used in conjunction from other findBiobankR package. It accepts an a data drame/data.table
with selected aliquots as a column
}
\examples{

library(data.table)
library(dplyr)
data("tb_data", package = "findBiobankR")
data("tb_specimen_df", package = "findBiobankR")

# required clinical cols
clinical_cols <- c("tb_group", "hiv_status", "sex", "country") 

clin_specs  <- convert_os_df_to_wide(clinical_df = tb_data,
                                     specimen_df = tb_specimen_df, 
                                     join_by_col = "ppid",
                                     specimen_col ="specimen_type" ,
                                     clinical_vars = clinical_cols ,
                                      specimen_label_col = "specimen_label",
                                     os_aliquot_names = c("Serum", "Plasma")
)


clin_specs[, serum_aliquot := select_one_aliquot(NP_labels_serum, aliquot_number = 1 )]
clin_specs[, plasma_aliquot := select_one_aliquot(NP_labels_plasma, aliquot_number= 1 )]

## prepare data set
selected_aliquots_df(clin_specs,
                     aliquot_cols = c("serum_aliquot","plasma_aliquot"),
                     ppid_col = "ppid")
}
\seealso{
findBiobank::select_one_aliquot
}
