% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_categorical_vars.R
\name{summarise_categorical_vars}
\alias{summarise_categorical_vars}
\title{Summarise categorical variables}
\usage{
summarise_categorical_vars(
  df,
  by_vec,
  names_from_var = NULL,
  setordr_col = NULL,
  names_from_sort = NULL,
  return_wide = TRUE,
  add_totals_where = c("row", "col")
)
}
\arguments{
\item{df}{data frame}

\item{by_vec}{The group by columns/variable}

\item{names_from_var}{dcast variable}

\item{setordr_col}{column to order}

\item{names_from_sort}{A vectors to sort columns created by pivot_wider}

\item{return_wide}{Logical vector whether to return a wide data.frame where names_from_var categories are column}

\item{add_totals_where}{where to add column sum totals check janitor::adorn_totals where argument}
}
\value{
a data.frame
}
\description{
Useful for for folks who work with a lot of data summaries. Produces a nice frequency table
}
\examples{
data("tb_data", package = "findBiobankR")
summarise_categorical_vars(df = tb_data,
                           by_vec = c("tb_group", "country"),
                           names_from_var ="country",
                           setordr_col = "tb_group",
                           names_from_sort =NULL ,
                           return_wide = TRUE,
                           add_totals_where = c("row", "col"))
}
