% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tb_data.R
\docType{data}
\name{tb_data}
\alias{tb_data}
\title{tb_data}
\format{
A data frame with 27 rows and 5 variables:
\describe{
\item{ ppid }{  character }
\item{ tb_group }{  character }
\item{ hiv_status }{  character }
\item{ sex }{  character }
\item{ country }{  factor }
}
}
\source{
Source
}
\usage{
tb_data
}
\description{
Description.
}
\keyword{datasets}
