% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tb_dictionary.R
\docType{data}
\name{tb_dictionary}
\alias{tb_dictionary}
\title{tb_dictionary}
\format{
A data frame with 4 rows and 2 variables:
\describe{
\item{ variable_name }{  character }
\item{ variable_description }{  character }
}
}
\source{
Source
}
\usage{
tb_dictionary
}
\description{
Description.
}
\keyword{datasets}
