% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tb_resp_symptoms.R
\docType{data}
\name{tb_resp_symptoms}
\alias{tb_resp_symptoms}
\title{tb_resp_symptoms}
\format{
A data frame with 5 rows and 6 variables:
\describe{
\item{ COUGH }{  character }
\item{ EXPECTORATION }{  character }
\item{ HEMOPTYSIS }{  character }
\item{ CHEST_PAIN }{  character }
\item{ DYSPNOE }{  character }
\item{ ID }{  integer }
}
}
\source{
Source
}
\usage{
tb_resp_symptoms
}
\description{
Description.
}
\keyword{datasets}
