% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_tb_specimen_df.R
\docType{data}
\name{tb_specimen_df}
\alias{tb_specimen_df}
\title{tb_specimen_df}
\format{
A data frame with 243 rows and 8 variables:
\describe{
\item{ ppid }{  character }
\item{ specimen_label }{  character }
\item{ specimen_type }{  factor }
\item{ visit_event }{  character }
\item{ specimen_requirement_name }{  factor }
\item{ specimen_barcode }{  character }
\item{ specimen_container_name }{  character }
\item{ specimen_container_position }{  integer }
}
}
\source{
Source
}
\usage{
tb_specimen_df
}
\description{
Description.
}
\keyword{datasets}
