% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_timestamp.R
\name{timestamp_to_date.data.frame}
\alias{timestamp_to_date.data.frame}
\title{Convert Timestamp Columns in a Data Frame to Date}
\usage{
\method{timestamp_to_date}{data.frame}(timestamp, date_cols = NULL, ...)
}
\arguments{
\item{timestamp}{Data frame with timestamp columns.}

\item{date_cols}{Names of the timestamp columns to convert.}

\item{...}{(Optional) Return the converted date as numeric. Default is FALSE.}
}
\value{
A modified data frame with converted timestamp columns.
}
\description{
This method converts timestamp columns in a data frame to date columns.
}
