% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp_to_date.R
\name{timestamp_to_date.numeric}
\alias{timestamp_to_date.numeric}
\title{Convert Numeric Timestamp to Date}
\usage{
\method{timestamp_to_date}{numeric}(timestamp, date_cols = NULL)
}
\arguments{
\item{timestamp}{Numeric timestamp.}

\item{date_cols}{Names of the timestamp columns to convert if the object timestamp is a data.frame.}
}
\value{
A datetime object.
}
\description{
This method converts a numeric timestamp to a datetime object.
}
