% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_sites.R
\name{get_all_sites}
\alias{get_all_sites}
\title{Get All Sites}
\usage{
get_all_sites(auth_response)
}
\arguments{
\item{auth_response}{Authentication response object obtained from the OpenSpecimen API.}
}
\value{
A data table containing the details of all sites, or NULL if an error occurs.
}
\description{
Use this function to retrieve all the sites in the OpenSpecimen application.
The function sends an HTTP GET request to the OpenSpecimen server to fetch
details of all the sites.
}
\note{
\href{https://openspecimen.atlassian.net/wiki/spaces/CAT/pages/1115685/Get+All+Sites}{see api docs for more details:}
}
\examples{
#get_all_sites()
}
\seealso{
\code{\link{auth_os}}, \code{\link{parse_os_response}}, \code{\link{parse_all_sites}}
}
