% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orders.R
\name{get_orders}
\alias{get_orders}
\title{Get OpenSpecimen Orders}
\usage{
get_orders(
  auth_response,
  include_stats = TRUE,
  start_at = NULL,
  max_results = NULL
)
}
\arguments{
\item{auth_response}{The authentication response obtained from \code{auth_os}.}

\item{include_stats}{Logical, whether to include statistics (default is TRUE).}

\item{start_at}{Integer, the starting position for pagination (optional).}

\item{max_results}{Integer, the maximum number of results to retrieve (optional).}
}
\value{
A data frame containing the parsed order data.
}
\description{
This function retrieves a list of orders from the OpenSpecimen API using the provided authentication response and optional query parameters.   https://openspecimen.atlassian.net/wiki/spaces/CAT/pages/259457077/Retrieve+list+of+Orders
}
\note{
This function allows you to retrieve orders from OpenSpecimen and parse the response data into a data frame.   https://openspecimen.atlassian.net/wiki/spaces/CAT/pages/259457077/Retrieve+list+of+Orders
}
\examples{
#get_orders()
}
