% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multiple_sheets.R
\name{read_multiple_sheets}
\alias{read_multiple_sheets}
\title{Read Sheets  From excel workbook}
\usage{
read_multiple_sheets(file_path, xl_sheets = NULL, ...)
}
\arguments{
\item{file_path}{path to file}

\item{xl_sheets}{excel sheets in the workbook}

\item{...}{Other parameters from read excel function}
}
\value{
returns a list of data frames

A list of data frames
}
\description{
Read Files From excel workbook
}
\examples{
iris_species <- system.file("extdata", "iris_species.xlsx",
                          package = "findBiobankR")

iris_list <- read_multiple_sheets(file_path  = iris_species ) \%>\% 
  lapply(head, 4)
iris_list
}
