% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_timestamp.R
\name{timestamp_to_date.numeric}
\alias{timestamp_to_date.numeric}
\title{Convert Timestamp to Date
This function converts a timestamp to a date time object}
\usage{
\method{timestamp_to_date}{numeric}(timestamp, date_cols = NULL, ...)
}
\arguments{
\item{timestamp}{The timestamp to convert.}

\item{date_cols}{(Optional) The columns in the data frame to convert.}

\item{...}{Additional arguments to be passed to the method.}
}
\value{
The converted date as a date time object.
}
\description{
Convert Timestamp to Date
This function converts a timestamp to a date time object
}
