% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp_to_date.R
\name{timestamp_to_date}
\alias{timestamp_to_date}
\title{Convert Timestamp to Date}
\usage{
timestamp_to_date(timestamp, date_cols = NULL)
}
\arguments{
\item{timestamp}{Numeric timestamp or a data frame with timestamp columns.}

\item{date_cols}{Names of the timestamp columns to convert.}
}
\value{
A datetime object when a numeric timestamp is provided, or a modified data frame when a
data frame with timestamp columns is provided.
}
\description{
This function provides a generic method and methods for numeric and data frame objects to convert
timestamps to date objects. It is especially useful for dealing with timestamps stored as numeric values.
}
\examples{

timestamp <- 1637892323000
datetime <- timestamp_to_date(timestamp)
datetime
# Example 2: Convert timestamp columns in a data frame
df <- data.frame(id = c(1, 2, 3),
                 event_date = c(1637892323000, 1637892423000, 1637892523000))


timestamp_to_date(df)

df
}
